refactor : PR 리뷰 후 게시글 라우터 및 서비스 함수 수정     1. foundAllBoard() -> findAllBoard() : 모든 게시글을 반환하는 서비스 함수 이름 변경
    2. 게시글 삭제 delete 라우터 및 서비스 함수 수정
     (1). 라우터와 서비스 함수의 역할 구분을 분명히 함
            접속한 사용자와 게시글 작성자와의 일치 여부를 판단하는 로직을 유틸 함수에서 서비스함수로 이동

    3. 게시글 서비스 함수
     (1). 게시글 조회 결과 0이 나오더라도 조회 성공이므로 수정
     (2). 정렬 방식에 따라 다른 시퀄라이즈 옵션을 할당하는 로직을 if..else문이 아닌 switch문을 사용하도록 수정
     (3). getBoardById() 서비스 함수의 반환 객체의 message 속성 값 수정 (페이지 -> 게시글)

+++ src/services/board.js
@@ -2,11 +2,11 @@ import { db } from '../models/index.js'
 import { Op } from 'sequelize'
 
 export class BoardService {
-    async foundAllBoard() {
+    async findAllBoard() {
         try {
             const allBoard = await db.Board.findAndCountAll() // 모델인스턴스의 배열 반환
 
-            if (allBoard.count > 0) {
+            if (allBoard.count >= 0) {
                 return {
                     result: true,
                     message: '게시글 조회 성공',
@@ -29,18 +29,31 @@ export class BoardService {
             const sort = queryDto.sort
             let sortOrder
 
-            if (sort === 'date_asc') sortOrder = ['createdAt', 'ASC']
-            else if (sort === 'date_desc') sortOrder = ['createdAt', 'DESC']
-            else if (sort === 'likes_asc') sortOrder = ['board_like', 'ASC']
-            else if (sort === 'likes_desc') sortOrder = ['board_like', 'DESC']
-            else if (sort === 'views_asc') sortOrder = ['board_view', 'ASC']
-            else if (sort === 'view_desc') sortOrder = ['board_view', 'DESC']
-            else {
-                return {
-                    result: false,
-                    message: '잘못된 정렬 방식',
-                    data: null,
-                }
+            switch (sort) {
+                case 'date_asc':
+                    sortOrder = ['createdAt', 'ASC']
+                    break
+                case 'date_desc':
+                    sortOrder = ['createdAt', 'DESC']
+                    break
+                case 'likes_asc':
+                    sortOrder = ['board_like', 'ASC']
+                    break
+                case 'likes_desc':
+                    sortOrder = ['board_like', 'DESC']
+                    break
+                case 'views_asc':
+                    sortOrder = ['board_view', 'ASC']
+                    break
+                case 'views_desc':
+                    sortOrder = ['board_view', 'DESC']
+                    break
+                default:
+                    return {
+                        result: false,
+                        message: '잘못된 정렬 방식',
+                        data: null,
+                    }
             }
 
             const searchedBoard = await db.Board.findAndCountAll({
@@ -52,7 +65,7 @@ export class BoardService {
                 order: [sortOrder], // 정렬 방식
             })
 
-            if (searchedBoard.count > 0) {
+            if (searchedBoard.count >= 0) {
                 return {
                     result: true,
                     message: '검색된 게시글 조회 성공',
@@ -101,8 +114,6 @@ export class BoardService {
             const original = await db.Board.findOne({
                 where: { board_id: boardId },
             }) // 반환객체 : Board{ dataValues: {}, _previousDataValues: {}, ... }
-
-            // console.log(original.dataValues)
 
             if (original instanceof db.Board) {
                 return {
@@ -164,19 +175,19 @@ export class BoardService {
             if (foundBoard instanceof db.Board) {
                 return {
                     result: true,
-                    message: '페이지 조회 성공',
+                    message: '게시글 조회 성공',
                     data: foundBoard,
                 }
             } else if (foundBoard === null) {
                 return {
                     result: false,
-                    message: '페이지 조회 실패',
+                    message: '게시글 조회 실패',
                     data: null,
                 }
             } else {
                 return {
                     result: false,
-                    message: '페이지 조회 오류',
+                    message: '게시글 조회 오류',
                     data: null,
                 }
             }
@@ -188,6 +199,14 @@ export class BoardService {
     async deleteBoardById(boardDto) {
         // 게시글 아이디 boardDto와 일치하는 게시글 데이터 삭제
         try {
+            const board = await db.Board.findOne({
+                where: { board_id: boardDto.boardId },
+            })
+
+            if (board.board_user_id !== boardDto.userId) {
+                return { result: false, message: '삭제 권한 없음' }
+            }
+
             const deletedRowNum = await db.Board.destroy({
                 where: { board_id: boardDto.boardId },
             }) // 반환값 : 삭제된 열의 갯수
